Distributed entry points to application modules.
