+++ setup.py
@@ -44,10 +44,10 @@ setuptools.setup(
         ],
     package_dir={'': 'source'},
     packages=setuptools.find_packages(where='source'),
-    entry_points={'console_scripts': ['fab=fab.entry:fab_entry',
-                                      'fab-grab=fab.entry:grab_entry',
-                                      'fab-dump=fab.entry:dump_entry'],
-                  'gui_scripts': ['fab-explorer=fab.entry:fab_explorer']},
+    entry_points={'console_scripts': ['fab=fab.builder:fab_entry',
+                                      'fab-grab=fab.grabber:grab_entry',
+                                      'fab-dump=fab.dumper:dump_entry'],
+                  'gui_scripts': ['fab-explorer=fab.explorer:explorer_entry']},
     python_requires='>=3.6, <4',
     install_requires=[],
     extras_require={
