--- unit-tests/entry_test.py
@@ -1,10 +1,0 @@
-##############################################################################
-# (c) Crown copyright Met Office. All rights reserved.
-# For further details please refer to the file COPYRIGHT
-# which you should have received as part of this distribution
-##############################################################################
-# The entry.py module holds the application entry points. By definition
-# testing on of these is effectively testing the entire application along with
-# command line handling and all manner of other complications.
-#
-# We defer such system testing to a set of specific system tests.
